+++ src/fides/api/schemas/connection_configuration/connection_config.py
@@ -13,6 +13,7 @@ from fides.api.schemas.connection_configuration import connection_secrets_schema
 from fides.api.schemas.policy import ActionType
 from fides.api.schemas.saas.saas_config import SaaSConfigBase
 from fides.api.util.connection_type import get_connection_type_secret_schema
+from fides.api.util.masking_util import mask_sensitive_fields
 
 
 class CreateConnectionConfiguration(BaseModel):
@@ -39,36 +40,6 @@ class CreateConnectionConfigurationWithSecrets(CreateConnectionConfiguration):
     secrets: Optional[connection_secrets_schemas] = None
     saas_connector_type: Optional[str] = None
     model_config = ConfigDict(from_attributes=True, extra="ignore")
-
-
-def mask_sensitive_fields(
-    connection_secrets: Dict[str, Any], secret_schema: Dict[str, Any]
-) -> Dict[str, Any]:
-    """
-    Mask sensitive fields in the given secrets based on the provided schema.
-    This function traverses the given secrets dictionary and uses the provided schema to
-    identify fields that have been marked as sensitive. The function replaces the sensitive
-    field values with a mask string ('********').
-    Args:
-        connection_secrets (Dict[str, Any]): The secrets to be masked.
-        secret_schema (Dict[str, Any]): The schema defining which fields are sensitive.
-    Returns:
-        Dict[str, Any]: The secrets dictionary with sensitive fields masked.
-    """
-    if connection_secrets is None:
-        return connection_secrets
-
-    secret_schema_properties = secret_schema["properties"]
-    new_connection_secrets = {}
-
-    for key, value in connection_secrets.items():
-        if key in secret_schema_properties:
-            if secret_schema_properties.get(key, {}).get("sensitive", False):
-                new_connection_secrets[key] = "**********"
-            else:
-                new_connection_secrets[key] = value
-
-    return new_connection_secrets
 
 
 class ConnectionConfigSecretsMixin(BaseModel):